@@ -1,27 +1,28 @@
 module.exports = {
-    extends: [
-        "airbnb",
-        "plugin:react/recommended",
-        "airbnb/base",
-        // "plugin:unicorn/recommended",
-        "prettier",
-        "prettier/prettier",
-        "plugin:prettier/recommended",
-    ],
-    env: {
-        browser: true,
-        es2021: true,
-    },
-    parserOptions: {
-        // ecmaFeatures: {
-        //     jsx: true,
-        // },
-        ecmaVersion: 12,
-        sourceType: "module",
-    },
-    plugins: ["react", "react-hooks"],
-    rules: {
-        "react/jsx-filename-extension": [1, { extensions: [".js", ".jsx"] }],
-        // your overrides
-    },
+    // extends: [
+    //     'airbnb',
+    //     'plugin:react/recommended',
+    //     'airbnb/base',
+    //     // "plugin:unicorn/recommended",
+    //     'prettier',
+    //     'prettier/prettier',
+    //     'plugin:prettier/recommended',
+    // ],
+    // env: {
+    //     browser: true,
+    //     es2021: true,
+    // },
+    // parserOptions: {
+    //     // ecmaFeatures: {
+    //     //     jsx: true,
+    //     // },
+    //     ecmaVersion: 12,
+    //     sourceType: 'module',
+    // },
+    // plugins: ['react', 'react-hooks'],
+    // rules: {
+    //     indent: 'off',
+    //     'react/jsx-filename-extension': [1, { extensions: ['.js', '.jsx'] }],
+    //     // your overrides
+    // },
 };
